# *: optimize query slow query table by avoid decode plan for each row