--- a/planner/core/rule_column_pruning.go
+++ b/planner/core/rule_column_pruning.go
@@ -287,8 +287,10 @@ func (p *LogicalMemTable) PruneColumns(parentUsedCols []*expression.Column) erro
 	switch p.TableInfo.Name.O {
 	case infoschema.TableStatementsSummary,
 		infoschema.TableStatementsSummaryHistory,
+		infoschema.TableSlowQuery,
 		infoschema.ClusterTableStatementsSummary,
-		infoschema.ClusterTableStatementsSummaryHistory:
+		infoschema.ClusterTableStatementsSummaryHistory,
+		infoschema.ClusterTableSlowLog:
 		// currently prune mem-table column only use for statements summary table.
 	default:
 		return nil