<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #25982

Special optimize the system table `slow_query`/`cluster_slow_query` query performance for `tidb-dashboard` to reduce query memory and time.

### Benchmark

Generate some to slow log, and each SQL length is `40960 byte`.

1. Lower the slow log threshold.
   ```sql
   set @@tidb_slow_log_threshold=0;
   ```
 
2. Use [loadgen](https://github.com/crazycs520/load) to generated load:

   ```shell
   bin/loadgen payload gen-stmt --host 172.16.5.40 --port 4019  --concurrency 50 --stmt-cnt=100 --query-len 40960
   ```
   The generated SQL is look like this:
   
   ```sql
   select * from t_gen_stmt_1 where a = 1 or b = 1 or c = 1 
   union all
   select * from t_gen_stmt_2 where a = 2 or b = 2 or c = 2 
   union all
   ...
   ...
   ...
   ```
   
   After generated about `3.6GB` slow log size, stop the `loadgen`. 

3. Enlarge the slow log threshold.
   ```sql
   set @@tidb_slow_log_threshold=3000;
   ```

4. Bench the SQL: 

```sql
SELECT Digest, Query, Conn_ID, (UNIX_TIMESTAMP(Time) + 0E0) AS timestamp, Query_time, Parse_time, 
	Compile_time, Mem_max, Process_time, Write_keys, Request_count 
FROM INFORMATION_SCHEMA.CLUSTER_SLOW_QUERY 
WHERE Time BETWEEN FROM_UNIXTIME(1626092537) AND FROM_UNIXTIME(1626180644 ) ORDER BY Query_time ASC LIMIT 100;
```

|                      | Master | This PR |
| ---------------------| ------ | ------- |
| Query duration       | 23s    | 8s      |
| Query memory consume | 3GB    | < 30MB  |

Related TiDB memory consume metric is below:

Master:

![image](https://user-images.githubusercontent.com/26020263/125459971-10eceb22-6ab4-4fa1-9a80-ad26acd5eb75.png)


This PR 

![image](https://user-images.githubusercontent.com/26020263/125460162-3805d114-8d71-4d77-a5a3-9de9f0587e51.png)


### What is changed and how it works?

By prune column and avoid `decode plan` when the query doesn't need the plan.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Manual test

Side effects

-N/A

Documentation

- N/A

### Release note <!-- bugfixes or new feature need a release note -->

- No release note